Rename package to 'structures'
