+++ src/structures/Geometry.py
